refactor: Making the template more dual-purpose
